fix: add util to reaschedule cron events, separate external schemas
