+++ wolfdesk/eventually/tickets/src/ticket.aggregate.ts
@@ -83,7 +83,10 @@ export const Ticket = (
         throw new errors.TicketIsNotOpenError(state.ticketId);
       if (state.closedById)
         throw new errors.TicketIsClosedError(state.ticketId);
-      // TODO: check if data.from/data.to match this ticket
+
+      // TODO: invariants
+      // only owners or agents can add new messages
+
       return Promise.resolve([
         bind("MessageAdded", { ...data, messageId: randomUUID() }),
       ]);
@@ -93,7 +96,10 @@ export const Ticket = (
         throw new errors.TicketIsNotOpenError(state.ticketId);
       if (state.closedById)
         throw new errors.TicketIsClosedError(state.ticketId);
-      // TODO: check other request escalation invariants
+
+      // TODO: invariants
+      // escalation can only be requested by owner after window expired
+
       return Promise.resolve([
         bind("TicketEscalationRequested", { ...data, requestId: randomUUID() }),
       ]);
@@ -103,7 +109,9 @@ export const Ticket = (
         throw new errors.TicketIsNotOpenError(state.ticketId);
       if (state.closedById)
         throw new errors.TicketIsClosedError(state.ticketId);
-      // TODO: check other escalation invariants
+
+      // TODO: invariants
+
       return Promise.resolve([
         bind("TicketEscalated", { ...data, escalationId: randomUUID() }),
       ]);
@@ -113,7 +121,9 @@ export const Ticket = (
         throw new errors.TicketIsNotOpenError(state.ticketId);
       if (state.closedById)
         throw new errors.TicketIsClosedError(state.ticketId);
-      // TODO: check other reassignment invariants
+
+      // TODO: invariants
+
       return Promise.resolve([bind("TicketReassigned", { ...data })]);
     },
     MarkMessageDelivered: (data, state) => {
@@ -132,6 +142,10 @@ export const Ticket = (
         throw new errors.TicketIsClosedError(state.ticketId);
       if (!state.messages[data.messageId])
         throw new errors.MessageNotFoundError(data.messageId);
+
+      // TODO: invariants
+      // message can only be acknowledged by receiver?
+
       return Promise.resolve([bind("MessageRead", { ...data })]);
     },
     MarkTicketResolved: (data, state) => {
@@ -139,7 +153,10 @@ export const Ticket = (
         throw new errors.TicketIsNotOpenError(state.ticketId);
       if (state.closedById)
         throw new errors.TicketIsClosedError(state.ticketId);
-      // TODO: check other resolve invariants
+
+      // TODO: invariants
+      // ticket can only be resolved by agent or owner?
+
       return Promise.resolve([bind("TicketResolved", { ...data })]);
     },
   },
